img-colorize/img-color-function/requirements.txt
azure-functions
numpy
opencv-python-headless
azure-storage-blob